[FEATURE JENKINS-8296] Allow use of random persona instead of a fix one
It would be great to allow per project use of a random persona at each launch of a build.

How :

In the configuration of a project, the "Random Persona" will be choose in the list of available personas.

At each launch of a build the plugin will verify if the choosen persona is the "Random Persona", if so this "Random Persona" will randomize the choose of one of the personas configured.

The remaining process will be the same as today.
The project quote and icon will of course reflect the last build.

Signed-off-by: Whren <whren@free.fr>